+++ plugins/blockly/runtume/fn/apply_first_match.ts
@@ -1,0 +1,9 @@
+import { Action } from "indicatorts";
+
+export function applyFirstMatch(actions1: Action[], actions2: Action[]) {
+  return actions1.map((action, index) => {
+    if (action === Action.HOLD && actions2[index] !== Action.HOLD)
+      return actions2[index];
+    return action;
+  });
+}
