+++ plugins/blockly/runtume/fn/index.ts
@@ -1,0 +1,6 @@
+import { applyFirstMatch } from "./apply_first_match";
+import { compare } from "./compare";
+import { booleanAlgebra } from "./boolean_algebra";
+import { cross } from "./cross";
+
+export default { applyFirstMatch, compare, booleanAlgebra, cross };
