+++ plugins/blockly/runtume/index.ts
@@ -1,0 +1,3 @@
+import fn from "./fn";
+
+export const runtime = { fn };
